# config/api.yaml
openapi: 3.0.0
info:
  title: Nazdikwala Api Docs
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development Server
  - url: https://dev.api.example.com/v1
    description: Production Server
...
components:
  securitySchemes:
    cookieToken:
      type: apiKey
      in: header
      name: nazdikwala
security:
  - cookieToken: []
paths:
...
  /super/admin/create/page:
    post:
      tags:
        - CMS
      summary: Create Page By Super Admin
      security:
        - cookieToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                pageTitle:
                  type: string
                  example: About Us
                pageContent:
                  type: string
                  example: About Us Nazdikwala
                isPublic:
                  type: boolean
                  example: true
                featuredImage:
                  type: string
                  format: binary
      responses:
        "201":
          description: Successful operation
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  # Get All Pages
  /pages:
    get:
      tags:
        - CMS
      summary: Get All Pages
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # Get Single Category By Id
  /page/{id}:
    get:
      tags:
        - CMS
      summary: Get Single Page
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the page to get
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  #Update Page Super Admin
  /super/admin/page/{id}:
    put:
      tags:
        - CMS
      summary: Update Page By (ID) Super Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the page to get
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                pageTitle:
                  type: string
                  example: About Us
                pageContent:
                  type: string
                  example: About Us Nazdikwala
                isPublic:
                  type: boolean
                  example: true
                featuredImage:
                  type: string
                  format: binary
                  description: If You Pass Empty Value Then This shopLogo Maybe Be Removed From DB
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

    # Delete Page Super Admin
    delete:
      tags:
        - CMS
      summary: Delete Page By (ID) Super Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the page to get
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  # -----------------------------------Coupons-------------------------------

  # Create Coupon Super Admin
  /super/admin/create/coupon:
    post:
      tags:
        - Coupons
      summary: Create Coupon By Super Admin
      security:
        - cookieToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                couponName:
                  type: string
                  example: Weekend
                couponCode:
                  type: string
                  example: JOY15
                discountPercent:
                  type: number
                  example: 15
                status:
                  type: string
                  example: active
                  enum: [active, inactive]
                  description: It Can Be "active" Or "inactive"
      responses:
        "201":
          description: Successful operation
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  # Create Coupon Admin
  /admin/create/coupon:
    post:
      tags:
        - Coupons
      summary: Create Coupon By Admin
      security:
        - cookieToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                couponName:
                  type: string
                  example: Weekend
                couponCode:
                  type: string
                  example: JOY15
                discountPercent:
                  type: number
                  example: 15
                status:
                  type: string
                  example: active
                  enum: [active, inactive]
                  description: It Can Be "active" Or "inactive"
      responses:
        "201":
          description: Successful operation
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  # Get All Coupons Super Admin
  /coupons:
    get:
      tags:
        - Coupons
      summary: Get All Coupons
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # Get Single Coupon By Id Super Admin
  /coupon/{id}:
    get:
      tags:
        - Coupons
      summary: Get Single Coupon
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the coupon to get
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # Get All Coupons Admin
  /admin/coupons:
    get:
      tags:
        - Coupons
      summary: Get All Coupons Admin
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  #Update Coupon Super Admin
  /super/admin/coupon/{id}:
    put:
      tags:
        - Coupons
      summary: Update Coupon By (ID) Super Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the coupon to get
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                couponName:
                  type: string
                  example: Weekend
                couponCode:
                  type: string
                  example: JOY15
                discountPercent:
                  type: number
                  example: 15
                status:
                  type: string
                  example: active
                  enum: [active, inactive]
                  description: It Can Be "active" Or "inactive"
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

    # Delete Coupon Super Admin
    delete:
      tags:
        - Coupons
      summary: Delete Coupon By (ID) Super Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the coupon to get
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  #Update Coupon Admin
  /admin/coupon/{id}:
    put:
      tags:
        - Coupons
      summary: Update Coupon By (ID) Super Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the coupon to get
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                couponName:
                  type: string
                  example: Weekend
                couponCode:
                  type: string
                  example: JOY15
                discountPercent:
                  type: number
                  example: 15
                status:
                  type: string
                  example: active
                  enum: [active, inactive]
                  description: It Can Be "active" Or "inactive"
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

    # Delete Coupon Admin
    delete:
      tags:
        - Coupons
      summary: Delete Coupon By (ID) Admin
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the coupon to get
      responses:
        "200":
          description: Successful operation
        "422":
          description: Not allowed
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error

  # -----------------------------------User-------------------------------

  # Register User
...